simple refactor
