+++ idunn/api/place.py
@@ -1,9 +1,36 @@
 from apistar import types, validators
+from idunn.blocks.base import BlocksValidator
+from idunn.blocks import PhoneBlock, OpeningHourBlock, InformationBlock, WebSiteBlock, ContactBlock
+
+LONG = "long"
+SHORT = "short"
+DEFAULT_VERBOSITY = LONG
+
+BLOCKS_BY_VERBOSITY = {
+    LONG: [
+        OpeningHourBlock,
+        PhoneBlock,
+        InformationBlock,
+        WebSiteBlock,
+        ContactBlock
+    ],
+    SHORT: [
+        OpeningHourBlock
+    ]
+}
 
 class Place(types.Type):
     PLACE_TYPE = ''
 
     type = validators.String()
+    id = validators.String(allow_null=True)
+    name = validators.String(allow_null=True)
+    local_name = validators.String(allow_null=True)
+    class_name = validators.String(allow_null=True)
+    subclass_name = validators.String(allow_null=True)
+    geometry = validators.Object(allow_null=True)
+    address = validators.Object(allow_null=True)
+    blocks = BlocksValidator(allowed_blocks=BLOCKS_BY_VERBOSITY.get(LONG))
 
     def __init__(self, *args, **kwargs):
         if not args:
@@ -18,50 +45,3 @@ class Place(types.Type):
     @classmethod
     def load_place(cls, es_place, lang, settings, verbosity):
         raise NotImplementedError
-
-class Admin(Place):
-    PLACE_TYPE = 'admin'
-
-    id = validators.String(allow_null=True)
-    name = validators.String(allow_null=True)
-    label = validators.String(allow_null=True)
-
-    @classmethod
-    def load_place(cls, es_place, lang, settings, verbosity):
-        return cls(
-            id=es_place.get('id', ''),
-            name=es_place.get('name', ''),
-            label=es_place.get('label', '')
-        )
-
-class Street(Place):
-    PLACE_TYPE = 'street'
-
-    id = validators.String(allow_null=True)
-    name = validators.String(allow_null=True)
-    label = validators.String(allow_null=True)
-
-    @classmethod
-    def load_place(cls, es_place, lang, settings, verbosity):
-        return cls(
-            id=es_place.get('id', ''),
-            name=es_place.get('name', ''),
-            label=es_place.get('label', '')
-        )
-
-class Address(Place):
-    PLACE_TYPE = 'address'
-
-    id = validators.String(allow_null=True)
-    name = validators.String(allow_null=True)
-    label = validators.String(allow_null=True)
-
-    @classmethod
-    def load_place(cls, es_place, lang, settings, verbosity):
-        return cls(
-            id=es_place.get('id', ''),
-            name=es_place.get('name', ''),
-            label=es_place.get('label', '')
-        )
-
-
