+++ idunn/utils/geom.py
@@ -1,0 +1,65 @@
+from idunn.api.place import BLOCKS_BY_VERBOSITY
+
+def build_blocks(es_poi, lang, verbosity):
+    blocks = []
+    for c in BLOCKS_BY_VERBOSITY.get(verbosity):
+        block = c.from_es(es_poi, lang)
+        if block is not None:
+            blocks.append(block)
+    return blocks
+
+def get_geom(es_poi):
+    """Return the correct geometry from the elastic response
+
+    A correct geometry means both lat and lon coordinates are required
+
+    >>> get_geom({}) is None
+    True
+
+    >>> get_geom({'coord':{"lon": None, "lat": 48.858260156496016}}) is None
+    True
+
+    >>> get_geom({'coord':{"lon": 2.2944990157640612, "lat": None}}) is None
+    True
+
+    >>> get_geom({'coord':{"lon": 2.2944990157640612, "lat": 48.858260156496016}})
+    {'coordinates': [2.2944990157640612, 48.858260156496016], 'center': [2.2944990157640612, 48.858260156496016], 'type': 'Point'}
+    """
+    geom = None
+    if 'coord' in es_poi:
+        coord = es_poi.get('coord')
+        lon = coord.get('lon')
+        lat = coord.get('lat')
+        if lon is not None and lat is not None:
+            geom = {
+                'coordinates': [lon, lat],
+                'center': [lon, lat],
+                'type': 'Point'
+            }
+    return geom
+
+def get_name(properties, lang):
+    """Return the Place name from the properties field of the elastic response
+    Here 'name' corresponds to the POI name in the language of the user request (i.e. 'name:{lang}' field).
+
+    If lang is None or if name:lang is not in the properties
+    Then name receives the local name value
+
+    'local_name' corresponds to the name in the language of the country where the POI is located.
+
+    >>> get_name({}, 'fr') is None
+    True
+
+    >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, None)
+    'spontini'
+
+    >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, 'cz')
+    'spontini'
+
+    >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, 'fr')
+    'spontinifr'
+    """
+    name = properties.get(f'name:{lang}')
+    if name is None:
+        name = properties.get('name')
+    return name
