+++ tests/test_full.py
@@ -41,6 +41,10 @@ def test_full():
                 2.3250037768187326,
                 48.86618482685007
             ],
+            "center": [
+                2.3250037768187326,
+                48.86618482685007
+            ],
             "type": "Point"
         },
         "address": {
